revise twisting column case
